+++ toolbox/kdtree_build.cpp
@@ -54,5 +54,5 @@ void mexFunction(int nlhs, mxArray * plhs[], int nrhs, const mxArray * prhs[]){
     // return the program a pointer to the created tree
     plhs[0] = mxCreateDoubleMatrix(1,1,mxREAL);
     double* pointer_to_tree = mxGetPr(plhs[0]);
-    pointer_to_tree[0] = (long) tree;
+    pointer_to_tree[0] = (intptr_t) tree;
 }
