+++ toolbox/kdtree_k_nearest_neighbors.cpp
@@ -9,8 +9,8 @@ void retrieve_tree( const mxArray* matptr, KDTree* & tree){
     // check that I actually received something
     if( pointer0 == NULL )
         mexErrMsgTxt("vararg{1} must be a valid k-D tree pointer\n");
-    // convert it to "long" datatype (good for addresses)
-    long pointer1 = (long) pointer0[0];
+    // convert it to "intptr_t" datatype (good for addresses)
+    intptr_t pointer1 = (intptr_t) pointer0[0];
     // convert it to "KDTree"
     tree = (KDTree*) pointer1;
     // check that I actually received something
